--- apps/charterafrica/src/pages/[slug]/[slug].spec.js
@@ -1,7 +1,0 @@
-const { test, expect } = require("@playwright/test");
-
-test("/about", async ({ page }) => {
-  await page.goto("/about");
-  const title = await page.title();
-  expect(title).toBe("About");
-});
