+++ apps/charterafrica/src/pages/index.spec.js
@@ -5,3 +5,9 @@ test("/", async ({ page }) => {
   const title = await page.title();
   expect(title).toBe("charter.AFRICA");
 });
+
+test("/about", async ({ page }) => {
+  await page.goto("/about");
+  const title = await page.title();
+  expect(title).toBe("About");
+});
