--- apps/charterafrica/src/pages/opportunities/[slug].snap.js
@@ -1,3 +1,0 @@
-// Jest Snapshot v1, https://goo.gl/fbAQLP
-
-exports[`/opportunities/[slug] renders unchanged 1`] = `<div />`;
