--- apps/charterafrica/src/pages/opportunities/[slug].test.js
@@ -1,21 +1,0 @@
-import { createRender } from "@commons-ui/testing-library";
-import React from "react";
-
-import Page from "./[slug].page";
-
-import theme from "@/charterafrica/theme";
-
-// eslint-disable-next-line testing-library/render-result-naming-convention
-const render = createRender({ theme });
-
-const defaultProps = {
-  locale: "en",
-  locales: ["en", "fr", "pt"],
-};
-
-describe("/opportunities/[slug]", () => {
-  it("renders unchanged", () => {
-    const { container } = render(<Page {...defaultProps} />);
-    expect(container).toMatchSnapshot();
-  });
-});
